pyproject.toml
[project]
name = "auditguard-mcp"
version = "0.1.0"
description = "Compliance-aware MCP server: PII safety, RBAC, policy enforcement, and structured audit logging for LLM tool use"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Rituraj", email = "rituraj@example.com" },
]
keywords = ["mcp", "pii", "compliance", "audit", "privacy-filter", "rbac"]

dependencies = [
    # Privacy Filter (local model, no API calls)
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "accelerate>=0.26.0",
    # MCP server
    "mcp[cli]>=1.0.0",
    # Schema validation
    "pydantic>=2.0.0",
    # Database and SQL parsing
    "sqlalchemy>=2.0.0",
    "sqlglot>=23.0.0",
    # Synthetic data
    "faker>=25.0.0",
    # Customer API server
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "httpx>=0.27.0",
    # LangGraph agent
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    # Environment variables
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
auditguard-mcp = "auditguard_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
